+++ src/main/io/flashfs.c
@@ -283,7 +283,7 @@ static void flashfsAdvanceTailInBuffer(uint32_t delta)
  *
  * Returns the number of bytes written
  */
-static void flashfsWriteCallback(uint32_t arg)
+static void flashfsWriteCallback(uintptr_t arg)
 {
     // Advance the cursor in the file system to match the bytes we wrote
     flashfsSetTailAddress(tailAddress + arg);
